// src/types/Datasheet.ts
export default interface Datasheet {
  id: string;
  name: string;
  faction_id: string;
  source_id: string;
  legend: string;
  role: string;
  loadout: string;
  transport: string;
  virtual: string;
  leader_head: string;
  leader_footer: string;
  damaged_w: string;
  damaged_description: string;
  link: string;
}